// src/app/Models/Account/User.ts
export class User {
  id?: number;
  userName?: string;
  lastName?: string;
  firstName?: string;
  birthDate?: Date;
  passwd?: string;
  email?: string;
  score?: number;
  petPreference?: string;
  Owner?: boolean;
  expirationClaims?: string;
  nbf?: number;
  exp?: number;
  iss?: string;
  aud?: string;
}